+++ packages/core/CHANGELOG.md
@@ -1,5 +1,75 @@
 # @crashlab/core
 
+## 0.1.15
+
+### Patch Changes
+
+- @crashlab/clock@0.1.15
+- @crashlab/filesystem@0.1.15
+- @crashlab/http-proxy@0.1.15
+- @crashlab/mongo@0.1.15
+- @crashlab/pg-mock@0.1.15
+- @crashlab/random@0.1.15
+- @crashlab/redis-mock@0.1.15
+- @crashlab/scheduler@0.1.15
+- @crashlab/tcp@0.1.15
+
+## 0.1.14
+
+### Patch Changes
+
+- @crashlab/clock@0.1.14
+- @crashlab/filesystem@0.1.14
+- @crashlab/http-proxy@0.1.14
+- @crashlab/mongo@0.1.14
+- @crashlab/pg-mock@0.1.14
+- @crashlab/random@0.1.14
+- @crashlab/redis-mock@0.1.14
+- @crashlab/scheduler@0.1.14
+- @crashlab/tcp@0.1.14
+
+## 0.1.13
+
+### Patch Changes
+
+- @crashlab/clock@0.1.13
+- @crashlab/filesystem@0.1.13
+- @crashlab/http-proxy@0.1.13
+- @crashlab/mongo@0.1.13
+- @crashlab/pg-mock@0.1.13
+- @crashlab/random@0.1.13
+- @crashlab/redis-mock@0.1.13
+- @crashlab/scheduler@0.1.13
+- @crashlab/tcp@0.1.13
+
+## 0.1.12
+
+### Patch Changes
+
+- @crashlab/clock@0.1.12
+- @crashlab/filesystem@0.1.12
+- @crashlab/http-proxy@0.1.12
+- @crashlab/mongo@0.1.12
+- @crashlab/pg-mock@0.1.12
+- @crashlab/random@0.1.12
+- @crashlab/redis-mock@0.1.12
+- @crashlab/scheduler@0.1.12
+- @crashlab/tcp@0.1.12
+
+## 0.1.11
+
+### Patch Changes
+
+- @crashlab/clock@0.1.11
+- @crashlab/filesystem@0.1.11
+- @crashlab/http-proxy@0.1.11
+- @crashlab/mongo@0.1.11
+- @crashlab/pg-mock@0.1.11
+- @crashlab/random@0.1.11
+- @crashlab/redis-mock@0.1.11
+- @crashlab/scheduler@0.1.11
+- @crashlab/tcp@0.1.11
+
 ## 0.1.10
 
 ### Patch Changes